Refactored undo_matrix by moving parts of its code to their own handler functions
To enhance readability. I also did some minor refactoring within `self._handle_unupdated_vectors` to remove one level of indent.

I also refactored the names of a few functions to make them private based on if they'll be used by other parts in the class and also how specific their use cases are.

I've noticed some bugs in this version of the app which I'll fix hopefully in the next few commits or so or less.