Created `general_utils.py`.
The new file contains a function that is a generalized version of the function `_set_empty_matrix_inputs_to_zero` located in `main.py`. The function is now named `set_nonetype_to_zero` and is located in the file made in this commit.

Why this was done:
A function with the exact same purpose as `_set_empty_matrix_inputs_to_zero` would have been used to fix the first bug detailed in the last commit, related to the `add_or_edit_vector` callback-function.
Instead of creating a brand new function with the same purpose as an already-existing function in the `main.py` file, a new file is made with the function that does the same thing as what the other two would have done.