Add basic implementation of applying an inverse.
Applying an inverse currently only handles the cases where an inverse matrix exists.

Attempting to apply a matrix to a non-invertible matrix will raise `NotImplementedError('WIP.')`.

Noted problem:
Applying an inverse-matrix of a matrix that has previously been applied the inverse-of will cause the first inverse matrix to have its values edited, without letting know the program that technically, another matrix was applied. 
This causes an issue where undoing all the matrices does not revert the graph back to a state where the basis vectors are unit length and in the right orientation.