Fix application of interpolated frames.
Basic animations done, only for the add_matrix function. Added relevant
documentation.

Current issues:
   1. The graph gets janky (lags) when you add multiple matrices before
the current animation has finished. This is top priority as every other
function will depend on this.
   2. The animations have only been applied to the add_matrix callback
function. No other function has this yet.
   3. The overall way each callback function should add frames to the
graph does not seem optimal right now, and in trying to fix the first
issue on this list, it would be redundant to apply the logic to every
function individually.