Merge #27: Revamp the Desktop Layout
Before this PR, the appearance of the web-app was unintuitive, not well-structured, and the graph itself was hard to see.
So, this PR gives the web-app a more modern layout that is more fit for presentation. Resolves #25.

This PR may be one of the last PRs before this repository becomes public. Though, this may not be the final layout--but its ease for modularity seems to be promising.

List of Commits Made:

* Separate layout from main file.

* Separate layout into different sections.

* Change old run_server() to run() for running app.

* Delete all styles and add bootstrap theme

* Split the layout into 4 files in a package.

* Add meta tags and fix website title.

* Move layout into tabs and delete unneeded elements

* Replace independent matrix-list updates with 1 callback

* Implement LaTeX to list matrices

* Add id to appropriate sections and divs with BEM

* Begin update on how graphs scale.

* Implement basic graph auto-scaling in animations

* (WIP) Add latest-matrix into matrix-section

* Lightly style matrix-section

* Rename matrix-section to fit naming conventions

* Add cosmetic brackets and make utility func for matrix-component

* Update styling for new matrix-component

* Patch minor bug in matrix-list handling.

* Change html.Button to dbc.Button for easy styling

* Replace text with undo-redo icons

* Move location of undo-redo buttons to save space.

* Add minor spacing

Just to make it temporarily look more pleasing.

* Reformat all files for consistency.

* Rework matrix_component and use it in vector-section

* Remove bad styling and adjust accordingly + rearrange vector sect.

* (WIP) Rework logs-section position and styling

Changes made:
- The logs section was moved to the bottom left corner of the screen
with a dbc. Accordion() item, and its styling was also updated to match.
- Logs are now displayed as an ordered list that goes upwards with every
new log.
- To do all this, the way the program adds new logs has been updated to
append to a list instead of a string, but this will be done in the next
commit.

* Update logs-section interactions with last commit

* Update undo-matrix to scale graph correctly.

* Make displaying matrices as a list, like logs-sect

* Deduplicate consecutive messages in log display

Current Problems:
- Redo matrix works differently for some reason due to a log from 2 days
ago? Need to get that checked.
- Undo matrix gives an error when undoing matrices successfully when the
log list is empty.

* Fix a bug in undo, redo, and repeat matrix

Undo matrix bug:
- Updating output-logs when there is nothing to be updated: Fix by
adding a None-check and returning no_update when no new logs are added.

Redo matrix bug:
- Accidentally overwrote an if statement: Fix by undoing that.
- Wrong log statement: Change the log to be more general to redoing
matrices.

Repeat matrix bug:
- Same as undo-matrix bug, which is updating when there are no new logs.
Simply replace the returned value with no_update.

* Add a length-check for empty lists in update_logs

This resolves any remaining "Failed to call update_output_logs_display"
errors caused by functions not returning "no_update" when the
output_logs storage is empty and said functions return an empty list.

* Lightly improve styling and hide unused tab in matrix-sect

* Apply graph scaling for all graph-changing funcs

* Reorder matrix entries for proper displaying